requirements.txt
openai-whisper
ffmpeg
ffmpeg-python
flask
pytube
transformers
requests